+++ crc32.h
@@ -1,4 +1,4 @@
 #pragma once
 
 // Computes the CRC32 of a buffer
-uint32_t crc32(void* buf, size_t len, uint32_t partial_crc = 0);
+uint32_t crc32(void* buf, uint32_t len, uint32_t partial_crc = 0);
